+++ tests/test_models/test_recognizers/test_recognizer2d.py
@@ -30,17 +30,32 @@ def test_tsn():
     for one_img in img_list:
         recognizer(one_img, gradcam=True)
 
-    # test mixup forward
-    config = get_recognizer_cfg(
-        'tsn/tsn_r50_video_mixup_1x1x8_100e_kinetics400_rgb.py')
-    config.model['backbone']['pretrained'] = None
+    mmcls_backbone = dict(
+        type='mmcls.ResNeXt',
+        depth=101,
+        num_stages=4,
+        out_indices=(3, ),
+        groups=32,
+        width_per_group=4,
+        style='pytorch')
+    config.model['backbone'] = mmcls_backbone
+
     recognizer = build_recognizer(config.model)
-    input_shape = (2, 8, 3, 32, 32)
+
+    input_shape = (1, 3, 3, 32, 32)
     demo_inputs = generate_recognizer_demo_inputs(input_shape)
+
     imgs = demo_inputs['imgs']
     gt_labels = demo_inputs['gt_labels']
+
     losses = recognizer(imgs, gt_labels)
     assert isinstance(losses, dict)
+
+    # Test forward test
+    with torch.no_grad():
+        img_list = [img[None, :] for img in imgs]
+        for one_img in img_list:
+            recognizer(one_img, None, return_loss=False)
 
 
 def test_tsm():
@@ -98,6 +113,7 @@ def test_tpn():
 
     losses = recognizer(imgs, gt_labels)
     assert isinstance(losses, dict)
+    assert 'loss_aux' in losses and 'loss_cls' in losses
 
     # Test forward test
     with torch.no_grad():